requirements.txt
beautifulsoup4==4.8.2
Django==2.2.5
python-vlc==3.0.7110